pyproject.toml
[project]
name = "variants"
version = '0.0.1'

dependencies = [
    "more-itertools",
    "scikit-allel",
    "pandas",
    "numpy",
    "pyarrow",
    "chunked_array_set",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

